Extracted FileTree part
