+++ src/helpers.ts
@@ -1,6 +1,7 @@
 import sortBy from 'lodash/sortBy'
 import cloneDeep from 'lodash/cloneDeep'
 import { IFileItem, ITreeItem } from './interfaces'
+import * as settings from './settings'
 
 export function exportDescriptor(descriptor: object) {
   const text = encodeURIComponent(JSON.stringify(descriptor, null, 2))
@@ -24,8 +25,11 @@ export function createFileTree(items: IFileItem[]): ITreeItem[] {
     const parts = item.path.split('/')
     const level = parts.length
     const name = parts[level - 1]
+    let type = 'file'
+    if (item.isFolder) type = 'folder'
+    if (item.path === settings.PACKAGE_PATH) type = 'package'
     maxLevel = Math.max(maxLevel, level)
-    tree.push({ name, path: item.path, children: [] })
+    tree.push({ name, path: item.path, type, children: [] })
   }
 
   // Unflatten tree
